### 2023-07-23 #5685 @arfon
@karthik - as a process improvement, what do you think about listing the rOpenSci reviewers (`@christopherkenny`, `@SebastianWojcik86`) as the reviewers on the JOSS paper too?
