### 2023-09-27 #5885 @editorialbot
[REVIEW]: concstats:  Market Structure, Concentration and Inequality Measures, with R
**Submitting author:** @schneiderpy (<a href="http://orcid.org/0000-0001-5630-1097">Andreas Schneider</a>)
**Repository:** https://github.com/ropensci/concstats
**Branch with paper.md** (empty if default branch):
**Version:** v.0.1.6
