### 2023-09-27 318c6de
Minor fixes to paper
