### 2023-09-27 #5885 @arfon
@schneiderpy, could you please merge this PR? https://github.com/ropensci/concstats/pull/2

Also, please check the possible missing DOI (`10.1504/ijaf.2009.029147 may be a valid DOI for title: Banking concentration and the price-concentration relationship: the case of Brazil`) and update your BibTeX file accordingly.
