+++ paper/paper.bib
@@ -45,7 +45,7 @@
   number = {4},
   pages = {415-435},
   month = {},
-  url = {https://www.inderscienceonline.com/doi/abs/10.1504/IJAF.2009.029147}
+  url = {https://doi.org/10.1504/IJAF.2009.029147}
 }
 
 @article{garcia:1994,
