### 2023-09-27 #5885 @schneiderpy
@arfon, thank you for you message. I have fixed the mentioned DOI and updated the bib. file accordingly. Furthermore, I have added the JOSS badge (Under review). Thank you.
