### 2023-09-28 #5885 @arfon
@schneiderpy – looks like the `\ref{fig:mstruct})` is throwing an error here. I think it might be the `:` in the label? Please check these docs and see if you can figure it out: https://joss.readthedocs.io/en/latest/submitting.html#internal-references
