### 2023-09-28 #5885 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
